requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
psycopg2-binary
pwdlib
pwdlib[argon2]
python-jose[cryptography]
httpx
psycopg[binary]
sqlalchemy
alembic
loguru
pytest
pytest-asyncio
pytest-mock
requests-mock
typing_extensions
google-genai
python-multipart